Move the sorting into its own separate utility class.
Create SuperTypesLastSorterTest to hold the sorting implementation. Adjust
the tests accordingly and use it in PropertyAndSetterInjection.
